+++ frontend/shared/interop-sample/domain/src/commonMain/kotlin/interopSample/models/Weather.kt
@@ -1,7 +1,9 @@
 package interopSample.models
 
+import utils.BigDecimal
+
 @JsExport
 data class Weather(
-    val temperature: Double,
+    val temperature: BigDecimal,
     val temperatureUnit: String
 )
